Rename make_yaml to trans_spec_gen, add CLI command and docs
- Rename src/dm_bip/make_yaml/ to src/dm_bip/trans_spec_gen/
- Add `dm-bip generate-trans-specs` Typer subcommand with --input,
  --output, --cohort, --entity, --template options
- Rewrite module README with usage, input CSV format, output structure
- Add trans_spec_gen to top-level README repo structure
