+++ src/dm_bip/cli.py
@@ -1,6 +1,7 @@
 """Command line interface for dm-bip."""
 
 import logging
+from pathlib import Path
 from typing import Annotated, Optional
 
 import typer
@@ -48,6 +49,33 @@ def run():
     """Display usage information for the dm-bip pipeline."""
     typer.echo("The dm-bip pipeline is run using make.")
     typer.echo("Run 'make help' to see available targets and usage information.")
+
+
+@app.command()
+def generate_trans_specs(
+    input_csv: Annotated[Path, typer.Option("--input", "-i", help="Path to the metadata CSV")],
+    output_dir: Annotated[Path, typer.Option("--output", "-o", help="Directory for YAML output files")],
+    cohort: Annotated[str, typer.Option("--cohort", "-c", help="Cohort to filter on (e.g. aric, jhs, whi)")],
+    entity: Annotated[str, typer.Option("--entity", "-e", help="Entity type to filter on")] = "MeasurementObservation",
+    template: Annotated[str, typer.Option("--template", "-t", help="Jinja2 template filename")] = "yaml_measobs.j2",
+):
+    """Generate trans-spec YAML files from a metadata CSV."""
+    from dm_bip.trans_spec_gen.generate_trans_specs import generate_yaml
+
+    results = generate_yaml(
+        input_csv=input_csv,
+        output_dir=output_dir,
+        entity=entity,
+        cohort=cohort,
+        template_name=template,
+    )
+    if results:
+        typer.echo(f"Generated {len(results)} YAML files in {output_dir}")
+        for path in results:
+            typer.echo(f"  {path}")
+    else:
+        typer.echo(f"No matching rows for entity={entity}, cohort={cohort}")
+        raise typer.Exit(code=1)
 
 
 if __name__ == "__main__":
