--- src/dm_bip/make_yaml/README.md
@@ -1,11 +1,0 @@
-# Make YAML
-
-Tools for converting Stata scripts (.do files) that generate YAML transformation specifications into equivalent Python scripts.
-
-## Status
-
-This module is under development as part of the DMC Integration work. See [GitHub Issue #175](https://github.com/linkml/dm-bip/issues/175) for progress on incorporating the Stata YAML authoring tool.
-
-## Purpose
-
-Some data teams use Stata scripts to generate LinkML-Map transformation specification YAML files. This module will provide Python equivalents or converters for those workflows.
