+++ src/dm_bip/trans_spec_gen/generate_trans_specs.py
