+++ src/dm_bip/trans_spec_gen/templates/yaml_measobs.j2
